made data package a subpackage of mapping called geodata
