--- webapp/server/src/utils/data/__init__.py
@@ -1,1 +1,0 @@
-"""This package contains files related to storing or loading different kinds of data."""
